requirements.txt
fastapi
uvicorn[standard]
gunicorn
httpx
python-dotenv
cryptography
sqlalchemy>=2.0
aiosqlite
mcp
python-jose[cryptography]
